+++ cypress/integration/Options.spec.ts
@@ -21,10 +21,11 @@ describe('Options', () => {
     cy.get('[data-testid=nickname-0] > input')
       .type('Bulba Nickname')
       .should('have.value', 'Bulba Nickname');
-    cy.get('#search-filter').click();
+    cy.get('[data-testid=filter] > input').click();
     cy.get('[data-testid=pokemon-1]').click();
     cy.get('[data-testid=poke-Bulbasaur]').click({ force: true });
     cy.contains('DUPE').should('exist');
+    cy.contains('Duplicate pokémon are not allowed').should('exist');
     cy.scrollTo('top');
     cy.contains('Starter').should('exist');
     cy.get('[data-testid=status-0]').click();
@@ -32,6 +33,8 @@ describe('Options', () => {
     cy.contains('Starter').should('not.exist');
     cy.get('[data-testid=nickname-1] > .ui').click();
     cy.get('[data-testid=nickname-1] > input').should('have.length.above', 0);
+    cy.get('[data-testid=search-options]').click();
+    cy.contains('Settings').should('exist');
   });
 
   it('About and Changelog', () => {
@@ -42,6 +45,9 @@ describe('Options', () => {
     cy.get('[data-testid=options]').click();
     cy.get('[data-testid=changelog]').click();
     cy.contains('Changelog').should('exist');
+    cy.get('[data-testid=changelog-list]').find('h4').should('have.length', 3);
+    cy.get('[data-testid=see-more]').click();
+    cy.get('[data-testid=changelog-list]').find('h4').should('have.length.above', 3);
     cy.contains('Back').click();
     cy.contains('!').should('not.exist');
   });
@@ -56,7 +62,7 @@ describe('Options', () => {
   });
 
   context('Import', () => {
-    it('Import - Success', () => {
+    it.only('Import - Success', () => {
       cy.get('[data-testid=import]').attachFile('PokemonList.json');
       cy.contains('Emerald Kaizo').should('exist');
       cy.contains('Super Secret Place').should('exist');
@@ -64,18 +70,15 @@ describe('Options', () => {
       cy.contains('Team').should('exist');
       cy.contains('Rules').click();
       cy.contains('Have fun!').should('exist');
+      cy.contains('Stats').click();
+      cy.contains('ONGOING').should('exist');
       cy.contains('Builder').click();
       cy.get('[data-testid=team-poke-Venusaur]').should('exist');
     });
 
     it('Import - Failure', () => {
-      cy.visit('/', {
-        onBeforeLoad(win) {
-          cy.stub(win.console, 'error').as('consoleError');
-        },
-      });
       cy.get('[data-testid=import]').attachFile('Invalid.json');
-      cy.get('@consoleError').should('be.calledOnce');
+      cy.contains('Invalid file').should('exist');
     });
   });